backend/requirements.txt
amqp==5.3.1
annotated-types==0.5.0
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.2.4
cachetools==4.2.2
celery==5.6.2
certifi==2022.9.24
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.27
django-celery==3.1.17
django-cors-headers==4.4.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-filter==25.1
django-js-asset==2.2.0
django-mptt==0.18.0
djangorestframework==3.16.1
environs==9.5.0
exceptiongroup==1.3.1
greenlet==3.0.1
humanize==4.8.0
idna==3.4
importlib_metadata==8.7.1
kombu==5.6.2
lxml==4.9.3
Markdown==3.9
marshmallow==3.19.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==21.3
Pillow==9.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.39
protobuf==4.21.12
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-stdnum==1.19
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
redis==7.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tornado==6.1
typing_extensions==4.8.0
tzdata==2025.2
tzlocal==4.2
urllib3==1.26.6
vine==5.1.0
wcwidth==0.2.9
xlrd==2.0.1
zipp==3.23.0
